Creates the package to simulate dynamics of a manipulator
The package gets a urdf model the manipulator and creates a KDL tree of it.
With the KDL tree the package simulates the manipulator dynamics, generating
joints accelerations and them joints velocities and positions by numerical integration.
Joints states are published in a ros topic and effort commands are received through
ros topic also.